requirements.txt
camelot-py==0.10.1
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
distro==1.6.0
et-xmlfile==1.1.0
ghostscript==0.7
numpy==1.22.1
opencv-python==4.5.5.62
openpyxl==3.0.9
pandas==1.4.0
pdfminer.six==20211012
pdftopng==0.2.3
pycparser==2.21
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
tabula-py==2.3.0
tabulate==0.8.9
